+++ changelog.md
@@ -1,3 +1,8 @@
+# 1.2.0
+- Add SkillTest for modular Tests
+  - default dc is now 4
+  - a single roll may return multiple successes for each multiple of the dc
+  - if you do not have the skill it will try using an unskilled attempt
 # 1.1.0
 - Add Settings for additional Skills
 # 1.0.1